demos/mcp-server/requirements.txt
fastmcp==2.12.3
psycopg-binary==3.2.10
uvicorn==0.36.0
